+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### [0.2.0] - 2025-03-18
+
 ### Added
 
 * Support for updating breakpoints and bookmarks after the buffer rewrite.
@@ -21,5 +23,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * `Tools > Pasfmt > Settings...` to configure various user-level options.
 * Format-on-save functionality.
 
-[Unreleased]: https://github.com/integrated-application-development/delphilint/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/integrated-application-development/delphilint/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/integrated-application-development/delphilint/compare/v0.1.0..v0.2.0
 [0.1.0]: https://github.com/integrated-application-development/delphilint/releases/tag/v0.1.0